move definitions from internal to src folder
