+++ test/internal/datatypes_test.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:c2bluetooth/c2bluetooth.dart';
 import 'package:c2bluetooth/enums.dart';
-import '../../lib/internal/datatypes.dart';
+import '../../lib/src/datatypes.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
